==> crop/farm/step5_MapSPAM_analysis.Rmd <==
---
title: "MapSPAM_analysis"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Script objective
Analyse MapSPAM data.
1. Produce a dataframe with total production for each crop and iso3c.
2. Measure proportion of SPAM production to FAOSTAT production.
3. Checking data for island nations without production

# Data description
FAOSTAT crop production (extracted: 10/15/2020)

# Preamble
```{r setup, include = FALSE}
# getting packages we want
library(here)
library(janitor)

# src script
source(here("_workflow/common.R"))
source(here("_workflow/common_spatial.R"))

# Raster templates
source(here("_spatial/template_raster.R"))

predatalayers_path <- "/home/shares/food-systems/Food_footprint/all_food_systems/predatalayers/"
``` 

# Produce a dataframe with total production for each crop and iso3c.
```{r}
prod <- 
  vroom::vroom(here("crop/farm/data/crop_codes_updated.csv")) %>% 
  select(split_rest_names) %>% 
  unique()

distinct_vars <- expand_grid(prod_system = c("_A","_H","_S","_I", "_L"),
                             split_rest_names   = unique(prod$split_rest_names))
prod_list <- c()

for(i in 1:nrow(distinct_vars)) {
  
  
  prod_list[[i]] <- 
    sprintf("/home/shares/food-systems/Food_footprint/all_food_systems/dataprep/crop/farm/scaled_maps_2017/crop_%s%s_scaled.tif",
            distinct_vars$split_rest_names[i],
            distinct_vars$prod_system[i])
  
}

prod_crop <- 
  lapply(prod_list, raster) %>% 
  stack()

prod_crop_rgns_2017 <- 
  zonal(prod_crop, 
        food_rgns_tif, 
        fun      = "sum", 
        progress = "text", 
        na.rm    = TRUE) %>% 
  as_tibble() %>% 
  rename(ID_0 = zone) %>% 
  pivot_longer(cols      = starts_with("crop"),
               names_to  = "split_rest_names", 
               values_to = "production") %>% 
  mutate(prod_system = case_when(str_detect(split_rest_names, "_A_") ~ "A",
                                 str_detect(split_rest_names, "_H_") ~ "H",
                                 str_detect(split_rest_names, "_I_") ~ "I",
                                 str_detect(split_rest_names, "_S_") ~ "S",
                                 str_detect(split_rest_names, "_L_") ~ "L")) %>% 
  mutate(split_rest_names = substr(split_rest_names, start = 6, stop = 9)) %>% 
  left_join(food_rgns)
```

# Integrating fodder crops into workflow.
```{r}
fodd_rainfed_scaled <- 
  raster(file.path(prep,
                   "crop/farm/scaled_maps_2017/crop_fodd_rainfed_scaled.tif",
                   fsep = ""))

fodd_irrigated_scaled <-
  raster(file.path(prep,
                   "crop/farm/scaled_maps_2017/crop_fodd_I_scaled.tif",
                   fsep = ""))

# National statistics (for Juliette)
fodd_rainfed_national <- 
  zonal(fodd_rainfed_scaled,
        food_rgns_tif, 
        fun      = "sum",  
        na.rm    = TRUE) %>% 
  as_tibble() %>% 
  rename(ID_0 = zone) %>% 
  left_join(food_rgns)
write_csv(fodd_rainfed_national,
          here("crop/farm/data/fodd_rainfed_national.csv"))

fodd_irrigated_national <- 
  zonal(fodd_irrigated_scaled,
        food_rgns_tif, 
        fun      = "sum",  
        na.rm    = TRUE) %>% 
  as_tibble() %>% 
  rename(ID_0 = zone) %>% 
  left_join(food_rgns)
write_csv(fodd_irrigated_national,
          here("crop/farm/data/fodd_irrigated_national.csv"))

# Calculate proportion of crop production that goes to H vs S/L in each country
rainfed_system_proportions <- 
  prod_crop_rgns_2017 %>% 
  filter(!prod_system %in% c("A", "I")) %>%
  mutate(prod_system = case_when(prod_system %in% c("L", "S") ~ "low_input",
                                 TRUE ~ prod_system)) %>% 
  group_by(iso3c, prod_system, Country, ID_0) %>% 
  summarise(production = sum(production, na.rm = TRUE)) %>% 
  ungroup(prod_system) %>% 
  mutate(proportion = production / sum(production, na.rm = TRUE)) %>% 
  select(-production) 

# Median or zero for gapfilling countries
median_H_prop <- 
  median(rainfed_system_proportions$proportion[rainfed_system_proportions$prod_system == "H"],
         na.rm = TRUE)

# Create table with proportion of production allocated to H vs L/S systems.
fodd_H_prop <- 
  zonal(fodd_rainfed_scaled,
        food_rgns_tif, 
        fun      = "sum",  
        na.rm    = TRUE) %>% 
  as_tibble() %>% 
  rename(ID_0 = zone) %>% 
  left_join(food_rgns) %>% 
  mutate(split_rest_names = "fodd",
         prod_system = "H") %>% 
  rename(rainfed_production = sum) %>% 
  left_join(rainfed_system_proportions) %>% 
  mutate(proportion = if_else(is.na(proportion) == TRUE,
                              median_H_prop,
                              proportion))

# Cell allocation map production
fodd_H_raster <- 
  fodd_H_prop %>% 
  left_join(food_rgns_xy, .) %>% 
  select(x, y, proportion) %>%
  rasterFromXYZ(crs = food_crs) %>% 
  overlay(.,
          fodd_rainfed_scaled,
          fun = function(r1, r2) {
            
            r1[is.na(r1)] <- median_H_prop
            r3 <- r1 * r2
            
            return(r3)
          })
plot(fodd_H_raster)

writeRaster(fodd_H_raster,
            file.path(prep,
                      "crop/farm/scaled_maps_2017/crop_fodd_H_scaled.tif",
                      fsep = ""),
            format = "GTiff", 
            overwrite = TRUE)

writeRaster(fodd_H_raster,
            file.path(predatalayers_path,
                      "fodd_H_production_2017.tif",
                      fsep = ""),
            format = "GTiff", 
            overwrite = TRUE)
```

# Integrate fodd within prod_crop_rgns_2017 df
```{r}
fodd_H <- 
  fodd_H_prop %>% 
  mutate(production = rainfed_production * proportion) %>% 
  select(-c(proportion, rainfed_production))

fodd_irrigated <- 
  zonal(fodd_irrigated_scaled,
        food_rgns_tif, 
        fun      = "sum", 
        progress = "text", 
        na.rm    = TRUE) %>% 
  as_tibble() %>% 
  rename(ID_0 = zone) %>% 
  left_join(food_rgns) %>% 
  mutate(split_rest_names = "fodd",
         prod_system = "I") %>% 
  rename(production  = sum)

prod_crop_rgns_2017_final <- 
  bind_rows(fodd_irrigated,
            fodd_H,
            prod_crop_rgns_2017)

write_csv(prod_crop_rgns_2017_final, 
          here("crop/farm/data/prod_crop_rgns_2017.csv"))
```


```{r}
# Calculate proportion that each crop contributes to nationally and globally
prod_prop <- 
  prod_crop_rgns_2017 %>%
  filter(!prod_system %in% "A") %>% 
  group_by(split_rest_names) %>% 
  mutate(prop = production / sum(production))

write_csv(prod_prop, here("crop/farm/data/prod_prop.csv"))
```


# Identify what crop/country combinations aren't accounted for in 2017
These do not have 2010 SPAM data so can't be scaled to 2017. (Issue #284)

```{r}
vroom(here("crop/farm/data/prod_prop.csv"), 
      col_select = c(split_rest_names,
                     tonnes_SPAM = "production",
                     iso3c)) %>% 
  group_by(iso3c, SPAM_short_name = split_rest_names) %>% 
  summarise(tonnes_SPAM = sum(tonnes_SPAM, na.rm = TRUE)) %>% 
  ungroup() %>%
  left_join(vroom(here("crop/farm/data/inf_values.csv")), .) %>% 
  view()
```
